pyproject.toml
[tool.poetry]
name = "linenotify"
version = "0.1.0"
description = ""
authors = ["Koutaro Mukai <mukai.k1011k@outlook.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
opencv-python = "^4.6.0.66"
requests = "^2.28.1"
Pillow = "^9.3.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.990"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
